+++ packages/account-connector/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-connector
 
+## 2.0.10-next.1
+
+### Patch Changes
+
+- Adding webauthn as direct dependency to connector package
+- Updated dependencies
+  - @sophon-labs/account-core@2.0.10-next.1
+
 ## 2.0.10-next.0
 
 ### Patch Changes